Added FinancierVMTest

+++ src/Financier.Adapter/BackupWriter.cs
@@ -7,62 +7,55 @@ using System.Linq;
 
 namespace Financier.Adapter
 {
-    public class BackupWriter : IDisposable
+    public class BackupWriter : IBackupWriter
     {
-        private readonly TextWriter _writer;
-        private readonly string _fileName;
-        private readonly Dictionary<string, List<string>> _entityColumnsOrder;
-
-        public BackupVersion BackupVersion { get; }
-
-        public BackupWriter(string fileName, BackupVersion backupVersion, Dictionary<string, List<string>> entityColumnsOrder)
+        public void GenerateBackup(
+            IEnumerable<Entity> entities,
+            string fileName,
+            BackupVersion backupVersion,
+            Dictionary<string, List<string>> entityColumnsOrder,
+            bool deleteRawFile = true)
         {
-            BackupVersion = backupVersion;
-            _entityColumnsOrder = entityColumnsOrder;
-            _fileName = fileName;
-            _writer = new StreamWriter(Path.GetFileNameWithoutExtension(_fileName));
-        }
+            using var writer = new StreamWriter(Path.GetFileNameWithoutExtension(fileName));
 
-        public void GenerateBackup(List<Entity> entities)
-        {
-            WriteHeader(_writer);
-            WriteBody(_writer, entities);
-            WriteFooter(_writer);
-            _writer.Flush();
-            _writer.Close();
-            var fileWithoutExt = Path.GetFileNameWithoutExtension(_fileName);
-            Compress(fileWithoutExt, _fileName);
-            if (File.Exists(fileWithoutExt))
+            WriteHeader(writer, backupVersion);
+            WriteBody(writer, entities, entityColumnsOrder);
+            WriteFooter(writer);
+            writer.Flush();
+            writer.Close();
+            var fileWithoutExt = Path.GetFileNameWithoutExtension(fileName);
+            Compress(fileWithoutExt, fileName);
+            if (deleteRawFile && File.Exists(fileWithoutExt))
             {
                 File.Delete(fileWithoutExt);
             }
         }
 
-        private void WriteHeader(TextWriter bw)
+        private void WriteHeader(TextWriter bw, BackupVersion backupVersion)
         {
-            bw.WriteLine($"{Backup.PACKAGE}:{BackupVersion.Package}");
-            bw.WriteLine($"{Backup.VERSION_CODE}:{BackupVersion.VersionCode}");
-            bw.WriteLine($"{Backup.VERSION_NAME}:{BackupVersion.Version}");
-            bw.WriteLine($"{Backup.DATABASE_VERSION}:{BackupVersion.DatabaseVersion++}");
+            bw.WriteLine($"{Backup.PACKAGE}:{backupVersion.Package}");
+            bw.WriteLine($"{Backup.VERSION_CODE}:{backupVersion.VersionCode}");
+            bw.WriteLine($"{Backup.VERSION_NAME}:{backupVersion.Version}");
+            bw.WriteLine($"{Backup.DATABASE_VERSION}:{backupVersion.DatabaseVersion++}");
             bw.WriteLine(Backup.START);
         }
 
-        private void WriteBody(TextWriter bw, List<Entity> entities)
+        private void WriteBody(TextWriter bw, IEnumerable<Entity> entities, Dictionary<string, List<string>> columnsOrder)
         {
-            ExportTable(bw, entities.OfType<Account>().ToArray());
-            ExportTable(bw, entities.OfType<AttributeDefinition>().ToArray());
-            ExportTable(bw, entities.OfType<CategoryAttribute>().ToArray());
-            ExportTable(bw, entities.OfType<TransactionAttribute>().ToArray());
-            ExportTable(bw, entities.OfType<Budget>().ToArray());
-            ExportTable(bw, entities.OfType<Category>().ToArray());
-            ExportTable(bw, entities.OfType<Currency>().ToArray());
-            ExportTable(bw, entities.OfType<Location>().ToArray());
-            ExportTable(bw, entities.OfType<Project>().ToArray());
-            ExportTable(bw, entities.OfType<Transaction>().ToArray());
-            ExportTable(bw, entities.OfType<Payee>().ToArray());
-            ExportTable(bw, entities.OfType<CCardClosingDate>().ToArray());
-            ExportTable(bw, entities.OfType<SmsTemplate>().ToArray());
-            ExportTable(bw, entities.OfType<CurrencyExchangeRate>().ToArray());
+            ExportTable(bw, entities.OfType<Account>(), columnsOrder);
+            ExportTable(bw, entities.OfType<AttributeDefinition>(), columnsOrder);
+            ExportTable(bw, entities.OfType<CategoryAttribute>(), columnsOrder);
+            ExportTable(bw, entities.OfType<TransactionAttribute>(), columnsOrder);
+            ExportTable(bw, entities.OfType<Budget>(), columnsOrder);
+            ExportTable(bw, entities.OfType<Category>(), columnsOrder);
+            ExportTable(bw, entities.OfType<Currency>(), columnsOrder);
+            ExportTable(bw, entities.OfType<Location>(), columnsOrder);
+            ExportTable(bw, entities.OfType<Project>(), columnsOrder);
+            ExportTable(bw, entities.OfType<Transaction>(), columnsOrder);
+            ExportTable(bw, entities.OfType<Payee>(), columnsOrder);
+            ExportTable(bw, entities.OfType<CCardClosingDate>(), columnsOrder);
+            ExportTable(bw, entities.OfType<SmsTemplate>(), columnsOrder);
+            ExportTable(bw, entities.OfType<CurrencyExchangeRate>(), columnsOrder);
         }
 
         private void WriteFooter(TextWriter bw)
@@ -70,25 +63,17 @@ namespace Financier.Adapter
             bw.Write(Backup.END);
         }
 
-        private void ExportTable<T>(TextWriter bw, T[] ent) where T : Entity
+        private void ExportTable<T>(TextWriter bw, IEnumerable<T> ent, Dictionary<string, List<string>> _entityColumnsOrder) where T : Entity
         {
-            if (ent.Length > 0)
+            foreach (var item in ent)
             {
-                foreach (var item in ent)
-                {
-                    bw.Write(item.ToBackupLines(_entityColumnsOrder));
-                }
+                bw.Write(item.ToBackupLines(_entityColumnsOrder));
             }
         }
 
         public static string GenerateFileName()
         {
             return DateTime.Now.ToString("yyyyMMdd'_'HHmmss'_'fff") + ".backup";
-        }
-
-        public void Dispose()
-        {
-            _writer.Dispose();
         }
 
         public void Compress(string sourceFile, string compressedFile)
